requirements.txt
requests
python-dotenv
python-telegram-bot>=21.0
Pillow
